Split loginprofile in two pages
