+++ src/pages/ProfilePage.stories.js
@@ -1,0 +1,12 @@
+import ProfilePage from './ProfilePage'
+
+export default {
+  title: 'ProfilePage',
+  component: ProfilePage,
+};
+
+const Template = args => <ProfilePage {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {};
+
